feat(zod): expose zod schemas for auth, clock, enums, and quote types
