+++ src/index.ts
@@ -11,91 +11,23 @@
 /  - Response Body: Typed as `xxxResponse`
 */
 
-/*
-/ Clock Endpoint Types
-*/
+import {
+  FiatType,
+  CryptoType,
+  TransferType,
+  FiatAccountType,
+  KycSchema,
+  FiatAccountSchema,
+  KycStatus,
+  TransferStatus,
+  WebhookEventType,
+  SupportedOperatorEnum,
+} from './enums'
 
-// Response body for GET /clock
-export type ClockResponse = {
-  time: string
-}
-
-/*
-/ Auth Endpoint Types
-*/
-
-// Request body for /auth/login endpoint
-export interface AuthRequestBody {
-  message: string
-  signature: string
-}
-
-/*
-/ Quote Endpoint Types
-*/
-
-// Request body for all /quote endpoints
-export type QuoteRequestBody = {
-  fiatType: FiatType
-  cryptoType: CryptoType
-  address: string
-  fiatAmount?: string
-  cryptoAmount?: string
-  country: string
-  region?: string
-}
-
-// Response body for all /quote endpoints
-export type QuoteResponse = {
-  quote: {
-    fiatType: FiatType
-    cryptoType: CryptoType
-    fiatAmount: string
-    cryptoAmount: string
-    guaranteedUntil: string
-    quoteId: string
-    transferType: TransferType
-    fee?: string
-    feeType?: FeeType
-    feeFrequency?: FeeFrequency
-  }
-  kyc: {
-    kycRequired: boolean
-    kycSchemas: QuoteResponseKycSchema[]
-  }
-  fiatAccount: Partial<Record<FiatAccountType, FiatAccountTypeQuoteData>>
-}
-
-// Helper type
-export type QuoteResponseKycSchema = {
-  kycSchema: KycSchema
-  allowedValues: {
-    [key: string]: string[]
-  }
-}
-
-export type QuoteErrorResponse = {
-  error: FiatConnectError
-  minimumFiatAmount?: string
-  maximumFiatAmount?: string
-  minimumCryptoAmount?: string
-  maximumCryptoAmount?: string
-}
-
-// Helper type
-export type FiatAccountTypeQuoteData = {
-  fiatAccountSchemas: QuoteResponseFiatAccountSchema[]
-  settlementTimeLowerBound?: string // ISO-8601 Duration
-  settlementTimeUpperBound?: string // ISO-8601 Duration
-}
-
-// Helper type
-export type QuoteResponseFiatAccountSchema = {
-  fiatAccountSchema: FiatAccountSchema
-  allowedValues: {
-    [key: string]: string[]
-  }
-}
+export * from './clock'
+export * from './auth'
+export * from './enums'
+export * from './quote'
 
 /*
 / KYC Endpoint Types
@@ -185,47 +117,6 @@ export type TransferStatusResponse = {
 
 export type FiatAccountId = string
 
-export enum TransferType {
-  TransferIn = 'TransferIn',
-  TransferOut = 'TransferOut',
-}
-
-export enum TransferStatus {
-  TransferStarted = 'TransferStarted',
-  TransferFiatFundsDebited = 'TransferFiatFundsDebited',
-  TransferSendingCryptoFunds = 'TransferSendingCryptoFunds',
-  TransferAmlFailed = 'TransferAmlFailed',
-  TransferReadyForUserToSendCryptoFunds = 'TransferReadyForUserToSendCryptoFunds',
-  TransferReceivedCryptoFunds = 'TransferReceivedCryptoFunds',
-  TransferComplete = 'TransferComplete',
-  TransferFailed = 'TransferFailed',
-}
-
-export enum KycStatus {
-  KycNotCreated = 'KycNotCreated',
-  KycPending = 'KycPending',
-  KycApproved = 'KycApproved',
-  KycDenied = 'KycDenied',
-  KycExpired = 'KycExpired',
-}
-
-export enum FeeType {
-  KycFee = 'KycFee',
-  PlatformFee = 'PlatformFee',
-}
-
-export enum FeeFrequency {
-  OneTime = 'OneTime',
-  Recurring = 'Recurring',
-}
-
-// Types for request bodies sent from FiatConnect webhooks
-export enum WebhookEventType {
-  KycStatusEvent = 'WebhookKycStatusEvent',
-  TransferInStatusEvent = 'WebhookTransferInStatusEvent',
-  TransferOutStatusEvent = 'WebhookTransferOutStatusEvent',
-}
-
 type WebhookEventPayload = {
   [WebhookEventType.KycStatusEvent]: {
     kycSchema: KycSchema
@@ -244,37 +135,6 @@ export type WebhookRequestBody<T extends WebhookEventType> = {
   payload: WebhookEventPayload[T]
 }
 
-// Errors returned by FiatConnect endpoints
-export enum FiatConnectError {
-  InvalidSignature = 'InvalidSignature',
-  GeoNotSupported = 'GeoNotSupported',
-  CryptoAmountTooLow = 'CryptoAmountTooLow',
-  CryptoAmountTooHigh = 'CryptoAmountTooHigh',
-  FiatAmountTooLow = 'FiatAmountTooLow',
-  FiatAmountTooHigh = 'FiatAmountTooHigh',
-  CryptoNotSupported = 'CryptoNotSupported',
-  FiatNotSupported = 'FiatNotSupported',
-  UnsupportedSchema = 'UnsupportedSchema',
-  InvalidSchema = 'InvalidSchema',
-  ResourceExists = 'ResourceExists',
-  ResourceNotFound = 'ResourceNotFound',
-  TransferNotAllowed = 'TransferNotAllowed',
-  KycExpired = 'KycExpired',
-  Unauthorized = 'Unauthorized',
-  SessionExpired = 'SessionExpired',
-  InvalidParameters = 'InvalidParameters',
-  ContractLoginNotSupported = 'ContractLoginNotSupported',
-  NonceInUse = 'NonceInUse',
-  IssuedTooEarly = 'IssuedTooEarly',
-  ExpirationTooLong = 'ExpirationTooLong',
-  InvalidFiatAccount = 'InvalidFiatAccount',
-}
-
-export enum Network {
-  Alfajores = 'Alfajores',
-  Mainnet = 'Mainnet',
-}
-
 /*
  * FiatConnect dynamic type definitions.
  *
@@ -285,63 +145,6 @@ export enum Network {
  * Any interfaces/enum values prefixed with `Mock` are not officially supported by the FiatConnect specification,
  * and are only included here as examples.
  */
-
-export enum FiatType {
-  USD = 'USD',
-  EUR = 'EUR',
-  BRL = 'BRL',
-  GNF = 'GNF',
-  INR = 'INR',
-  NGN = 'NGN',
-  GHS = 'GHS',
-  KES = 'KES',
-  ZAR = 'ZAR',
-  PHP = 'PHP',
-  UGX = 'UGX',
-  GBP = 'GBP',
-  XOF = 'XOF',
-  RWF = 'RWF',
-  CNY = 'CNY',
-  XAF = 'XAF',
-  ARS = 'ARS',
-  BOB = 'BOB',
-  CLP = 'CLP',
-  COP = 'COP',
-  FKP = 'FKP',
-  GYD = 'GYD',
-  PYG = 'PYG',
-  PEN = 'PEN',
-  SRD = 'SRD',
-  UYU = 'UYU',
-  VES = 'VES',
-  MXN = 'MXN',
-  PAB = 'PAB',
-}
-
-export enum CryptoType {
-  cUSD = 'cUSD',
-  cEUR = 'cEUR',
-  cREAL = 'cREAL',
-  CELO = 'CELO',
-}
-
-export enum KycSchema {
-  PersonalDataAndDocuments = 'PersonalDataAndDocuments',
-}
-
-export enum FiatAccountSchema {
-  AccountNumber = 'AccountNumber',
-  MobileMoney = 'MobileMoney',
-  DuniaWallet = 'DuniaWallet',
-  IBANNumber = 'IBANNumber',
-  IFSCAccount = 'IFSCAccount',
-}
-
-export enum FiatAccountType {
-  BankAccount = 'BankAccount',
-  MobileMoney = 'MobileMoney',
-  DuniaWallet = 'DuniaWallet',
-}
 
 interface RequiredFiatAccountSchemaFields {
   accountName: string
@@ -378,13 +181,6 @@ type IFSCAccount = RequiredFiatAccountSchemaFields & {
   accountNumber: string
   country: string
   fiatAccountType: FiatAccountType.BankAccount
-}
-
-export enum SupportedOperatorEnum {
-  ORANGE = 'ORANGE',
-  MOOV = 'MOOV',
-  MTN = 'MTN',
-  WAVE = 'WAVE',
 }
 
 // Map of all supported fiat account schemas to the corresponding schema type. List must be manually updated
